+++ database/seedDatabase.js
@@ -1,5 +1,6 @@
 const mongoose = require('mongoose');
 const faker = require('faker');
+const Photo = require('./Model.js')
 
 mongoose.connect('mongodb://localhost/photos');
 
@@ -8,14 +9,8 @@ const db = mongoose.connection;
 db.on('err', console.error.bind('err'));
 db.once('open', () => {console.log('DB Connected')});
 
-var Schema = new mongoose.Schema({
-    listingId: Number,
-    photoId: Number,
-    caption: String,
-    photoUrl: String
-});
 
-var Photo = mongoose.model('Photos', Schema);
+
 
 var data = [];
 for (var i = 0; i < 10; i++) {
