+++ package.json
@@ -18,6 +18,7 @@
     "webpack": "^2.2.1"
   },
   "dependencies": {
+    "axios": "^0.19.0",
     "express": "^4.15.0",
     "faker": "^4.1.0",
     "mongoose": "^5.7.4",
